requirements.txt
alerta<6.0.0
cryptography>=2.2.1
cssselect==1.0.1 # BSD License
elasticsearch  # Apache-2.0
ipaddress>=1.0.17
keystoneauth1>=3.8.0
kubernetes
lxml>=3.7.2 # BSD License
oslo.serialization!=2.19.1,>=2.18.0
oslo.utils>=3.33.0
paramiko==2.0.0 # LGPLv2.1+
pbr!=2.1.0,>=2.0.0
pika>=0.10.0 # BSD License
pyOpenSSL>=17.1.0
pymongo
pytest<4.1.0,>=3.0.4  # MIT
pytest-html
pytest-ordering
pytest-rerunfailures
python-ceilometerclient>=2.5.0  # Apache-2.0
python-cinderclient>=3.3.0  # Apache-2.0
python-dateutil>=2.5.3
python-glanceclient>=2.8.0  # Apache-2.0
python-heatclient>=1.6.1  # Apache-2.0
python-keycloak
python-keystoneclient>=3.17.0  # Apache-2.0
python-muranoclient>=0.8.2  # Apache-2.0
python-neutronclient>=5.1.0  # Apache-2.0
python-novaclient>=9.1.0  # Apache-2.0
python-saharaclient>=1.1.0  # Apache-2.0
PyYAML>=3.12,<4  # MIT
requests>=2.14.2
six>=1.11.0  # MIT
tox>=2.5.0  # MIT
urllib3<1.24,>=1.21.1
warlock<2,>=1.2.0
